- SubSnpCoreFieldsRowMapper: subsnp id is now long instead of int - Added test data

+++ dbsnp-importer/src/main/java/uk/ac/ebi/eva/dbsnpimporter/io/readers/SubSnpCoreFieldsRowMapper.java
@@ -62,7 +62,7 @@ public class SubSnpCoreFieldsRowMapper implements RowMapper<SubSnpCoreFields> {
     @Override
     public SubSnpCoreFields mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         return new SubSnpCoreFields(
-                resultSet.getInt(SUBSNP_ID_COLUMN),
+                resultSet.getLong(SUBSNP_ID_COLUMN),
                 resultSet.getObject(REFSNP_ID_COLUMN, Long.class),
                 resultSet.getInt(SNP_ORIENTATION_COLUMN),
                 resultSet.getString(CONTIG_NAME_COLUMN),
